back/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.6
Flask-JWT-Extended==4.7.0
Flask-CORS==4.0.0
psycopg[binary]==3.1.18
python-dotenv==1.0.0
Werkzeug==3.0.1